Extract scope into its own class.
Asking the “Client” to gate or validate scopes seemed confusing, since
the scopes are specified on the token in OAuth, rather than on the
Client. (Although client’s can limit the scopes that are granted to a
token!)
