+++ app/Models/Client.php
@@ -4,7 +4,6 @@ namespace Northstar\Models;
 
 use Illuminate\Support\Str;
 use Jenssegers\Mongodb\Model;
-use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 /**
  * The Client model. These identify the "client application" making
@@ -50,19 +49,6 @@ class Client extends Model
      */
     protected $attributes = [
         'scope' => [],
-    ];
-
-    /**
-     * Available API Key scopes.
-     * @var array
-     */
-    protected static $scopes = [
-        'admin' => [
-            'description' => 'Allows "administrative" actions that should not be user-accessible, like deleting user records.',
-        ],
-        'user' => [
-            'description' => 'Allows actions to be made on a user\'s behalf.',
-        ],
     ];
 
     /**
@@ -150,67 +136,14 @@ class Client extends Model
     }
 
     /**
-     * Validate if all the given scopes are valid.
-     *
-     * @param $scopes
-     * @return bool
-     */
-    public static function validateScopes($scopes)
-    {
-        if (! is_array($scopes)) {
-            return false;
-        }
-
-        return ! array_diff($scopes, array_keys(static::$scopes));
-    }
-
-    /**
-     * Return a list of all scopes & their descriptions.
-     *
-     * @return array
-     */
-    public static function scopes()
-    {
-        return static::$scopes;
-    }
-
-    /**
      * Get the API key specified for the current request.
      *
-     * @return \Northstar\Models\ApiKey
+     * @return \Northstar\Models\Client
      */
     public static function current()
     {
         $api_key = request()->header('X-DS-REST-API-Key');
 
         return static::where('api_key', $api_key)->first();
-    }
-
-    /**
-     * Return whether a properly scoped API key is provided
-     * with the current request.
-     *
-     * @param $scope - Required scope
-     * @return bool
-     */
-    public static function allows($scope)
-    {
-        $key = self::current();
-
-        return $key && $key->hasScope($scope);
-    }
-
-    /**
-     * Throw an exception if a properly scoped API key is not
-     * provided with the current request.
-     *
-     * @param $scope - Required scope
-     */
-    public static function gate($scope)
-    {
-        if (! static::allows($scope)) {
-            app('stathat')->ezCount('invalid API key error');
-            throw new AccessDeniedHttpException('You must be using an API key with "'.$scope.'" scope to do that.');
-        }
     }
 }
