+++ app/Providers/ValidationServiceProvider.php
@@ -3,7 +3,7 @@
 namespace Northstar\Providers;
 
 use Illuminate\Support\ServiceProvider;
-use Northstar\Models\Client;
+use Northstar\Auth\Scope;
 
 class ValidationServiceProvider extends ServiceProvider
 {
@@ -25,7 +25,7 @@ class ValidationServiceProvider extends ServiceProvider
 
         // Add custom validator for localized date (e.g. `MM/DD/YYYY` or `DD/MM/YYYY`).
         $this->validator->extend('scope', function ($attribute, $value, $parameters) {
-            return Client::validateScopes($value);
+            return Scope::validateScopes($value);
         }, 'Invalid scope(s) provided.');
     }
